Refactored test suites
