+++ src/test/scala/zio/nio/BufferSpec.scala
@@ -11,17 +11,16 @@ import java.nio.{
   ShortBuffer => JShortBuffer
 }
 
-import zio.nio.channels.ZIOBaseSpec
+import scala.reflect.ClassTag
+
 import zio.{ Chunk, IO }
 import zio.test._
 import zio.test.Assertion._
-
-import scala.reflect.ClassTag
-import BufferSpecUtils._
 import zio.test.mock.MockEnvironment
+import BufferSpecUtils._
 
 object BufferSpec
-    extends ZIOBaseSpec(
+    extends BaseSpec(
       suite("BufferSpec")(
         commonBufferTests(
           "ByteBuffer",
@@ -92,39 +91,43 @@ object BufferSpecUtils {
     def zeroValues = Array(0, 0, 0).map(f)
 
     suite(suiteName)(
-      testM[MockEnvironment, Exception, String]("apply") {
+      testM("apply") {
         for {
-          apply <- allocate(3).flatMap(_.array)
-        } yield assert(apply.sameElements(zeroValues), isTrue)
+          allocated <- allocate(3)
+          array     <- allocated.array
+        } yield assert(array.sameElements(zeroValues), isTrue)
       },
-      testM[MockEnvironment, Exception, String]("wrap backed by an array") {
+      testM("wrap backed by an array") {
         for {
           buffer <- wrap(Chunk.fromArray(initialValues))
           array  <- buffer.array
         } yield assert(array.sameElements(initialValues), isTrue)
       },
-      testM[MockEnvironment, Exception, String]("capacity") {
+      testM("capacity") {
         for {
-          capacity <- allocate(initialCapacity).map(_.capacity)
+          allocated <- allocate(initialCapacity)
+          capacity  = allocated.capacity
         } yield assert(capacity == jAllocate(initialCapacity).capacity, isTrue)
       },
-      testM[MockEnvironment, Exception, String]("capacity initialized") {
+      testM("capacity initialized") {
         for {
-          capacity <- allocate(initialCapacity).map(_.capacity)
+          allocated <- allocate(initialCapacity)
+          capacity  = allocated.capacity
         } yield assert(capacity == initialCapacity, isTrue)
       },
-      testM[MockEnvironment, Exception, String]("position is 0") {
+      testM("position is 0") {
         for {
-          position <- allocate(initialCapacity).flatMap(_.position)
+          allocated <- allocate(initialCapacity)
+          position  <- allocated.position
         } yield assert(position == 0, isTrue)
       },
-      testM[MockEnvironment, Exception, String]("limit is capacity") {
+      testM("limit is capacity") {
         for {
-          limit <- allocate(initialCapacity).flatMap(_.limit)
+          allocated <- allocate(initialCapacity)
+          limit     <- allocated.limit
         } yield assert(limit == initialCapacity, isTrue)
       },
-      testM[MockEnvironment, Exception, String]("position set") {
-
+      testM("position set") {
         for {
           buffer   <- Buffer.byte(initialCapacity)
           _        <- buffer.position(3)
@@ -141,7 +144,6 @@ object BufferSpecUtils {
         } yield assert(newLimit == limit, isTrue)
       },
       testM("position reset") {
-
         for {
           buffer   <- Buffer.byte(initialCapacity)
           newLimit = 3
@@ -152,7 +154,6 @@ object BufferSpecUtils {
 
       },
       testM("reset to marked position") {
-
         for {
           b           <- allocate(initialCapacity)
           _           <- b.position(1)
@@ -164,7 +165,6 @@ object BufferSpecUtils {
 
       },
       testM("clear") {
-
         for {
           b        <- allocate(initialCapacity)
           _        <- b.position(1)
@@ -173,7 +173,6 @@ object BufferSpecUtils {
           position <- b.position
           limit    <- b.limit
         } yield assert(position == 0 && limit == initialCapacity, isTrue)
-
       },
       testM("flip") {
         for {
@@ -183,7 +182,6 @@ object BufferSpecUtils {
           position <- b.position
           limit    <- b.limit
         } yield assert(position == 0 && limit == 1, isTrue)
-
       },
       testM("rewind sets position to 0") {
         for {
@@ -191,18 +189,17 @@ object BufferSpecUtils {
           _           <- b.position(1)
           _           <- b.rewind
           newPosition <- b.position
-        } yield assert(newPosition == 1, isTrue)
+        } yield assert(newPosition == 0, isTrue)
       },
       testM("heap buffers a backed by an array") {
         for {
           b <- allocate(initialCapacity)
         } yield assert(b.hasArray, isTrue)
-
       },
       testM[MockEnvironment, Exception, String]("0 <= mark <= position <= limit <= capacity") {
-        checkM(Gen.anyInt, Gen.anyInt, Gen.anyInt, Gen.anyInt) {
+        checkM(Gen.int(-1, 10), Gen.int(-1, 10), Gen.int(-1, 10), Gen.int(-1, 10)) {
           (markedPosition: Int, position: Int, limit: Int, capacity: Int) =>
-            for {
+            (for {
               b    <- Buffer.byte(capacity)
               _    <- b.limit(limit)
               _    <- b.position(markedPosition)
@@ -210,7 +207,11 @@ object BufferSpecUtils {
               _    <- b.position(position)
               _    <- b.reset
               mark <- b.position
-            } yield assert(0 <= mark && mark <= position && position <= limit && limit <= capacity, isTrue)
+            } yield assert(0 <= mark && mark <= position && position <= limit && limit <= capacity, isTrue))
+              .catchSome {
+                case _: IllegalArgumentException | _: IllegalStateException =>
+                  IO.effectTotal(assert(true, isTrue))
+              }
         }
       }
     )
